requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
idna==2.6
numpy==1.14.4
pytz==2018.4
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
sklearn==0.0
urllib3==1.22
